pyproject.toml
[project]
name = "zsynctech-studio-sdk"
version = "0.2.0"
description = "Python SDK for building robots that integrate with the ZSyncTech Studio platform"
readme = "README.md"
authors = [
    { name = "Rodrigo Zavan", email = "rodrigo.zavan@zsynctech.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.scripts]
zsynctech-studio-sdk = "zsynctech_studio_sdk:main"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"
